fix(pipeline): address CI gate findings surfaced by deploy.sh test
- clippy: drop unnecessary u64 casts (np2p storage, media replication requeue)
- eslint: DuplicatesLightbox reset/clamp effects -> render-phase state
  adjustment (react-hooks/set-state-in-effect); PersonStore let->const
- dev compose: correct ai-server build context (./ai -> repo root) so the
  dev AI image can be rebuilt in sync with current gRPC health contract

All surfaced by running the new deploy.sh test stage (cargo check/clippy/
unit, full integration suite, client lint/test/build).

+++ np2p/src/storage/disk.rs
@@ -27,7 +27,7 @@ fn get_available_space(path: &Path) -> std::io::Result<u64> {
     unsafe {
         let mut stats: libc::statvfs = std::mem::zeroed();
         if libc::statvfs(c_path.as_ptr(), &mut stats) == 0 {
-            Ok(stats.f_frsize as u64 * stats.f_bavail as u64)
+            Ok(stats.f_frsize * stats.f_bavail)
         } else {
             Err(std::io::Error::last_os_error())
         }
